third_party/blink/web_tests/fast/inline/middle-continuation-inherits-visibility-from-inline-parent.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
.hidden {
  visibility:hidden;
}
</style>
<a href="#" class="hidden">
  <div id="div">Inside a</div>
</a>
<script>
test(function() {
  if (window.eventSender) {
    eventSender.dragMode = false;
    eventSender.mouseMoveTo(10, 10);
    var originalInfo = internals.getCurrentCursorInfo();
    assert_equals(originalInfo, "type=Pointer");
  }
}, 'crbug.com/706324: Middle continuations should inherit the style of their inline parent, so when the cursor is placed over the anonymous block it should remain a pointer.');
</script>
